migrate build output parsing function in build.py
